complete the creation of forgot and reset password methods
