+++ src/Controller/Security/AuthenticationController.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace App\Controller\Security;
+
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Annotation\Route;
+
+
+class AuthenticationController extends AbstractController
+{
+    /**
+     * @param Request $request
+     * @return Response
+     * @Route("/connexion", name="security_login")
+     */
+    public function login()
+    {
+        return $this->render('security/login.html.twig');
+    }
+
+    /**
+     * @Route("/deconnexion", name="security_logout")
+     */
+    public function logout() {}
+
+}
