+++ src/Controller/Security/RegistrationController.php
@@ -1,0 +1,106 @@
+<?php
+
+namespace App\Controller\Security;
+
+use App\Entity\User;
+use App\Form\RegistrationType;
+use App\Repository\UserRepository;
+use App\Service\Mailer;
+
+use Doctrine\ORM\EntityManagerInterface;
+
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Mailer\MailerInterface;
+use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
+use Symfony\Component\Security\Csrf\TokenGenerator\TokenGeneratorInterface;
+
+class RegistrationController extends AbstractController
+{
+    /**
+     * @var MailerInterface
+     */
+    private $mailer;
+    /**
+     * @var UserRepository
+     */
+    private $userRepository;
+
+    /**
+     * SecurityController constructor.
+     * @param Mailer $mailer
+     * @param UserRepository $userRepository
+     */
+    public function __construct(Mailer $mailer, UserRepository $userRepository)
+    {
+        $this->mailer = $mailer;
+        $this->userRepository = $userRepository;
+    }
+
+    /**
+     * @Route("/inscription", name="security_registration")
+     * @param Request $request
+     * @param EntityManagerInterface $manager
+     * @param UserPasswordEncoderInterface $encoder
+     * @param TokenGeneratorInterface $tokenGenerator
+     * @return Response
+     */
+    public function registration(Request $request, EntityManagerInterface $manager, UserPasswordEncoderInterface $encoder, TokenGeneratorInterface $tokenGenerator)
+    {
+        $user = new User;
+        $form = $this->createForm(RegistrationType::class, $user);
+
+        $form->handleRequest($request);
+
+        if($form->isSubmitted() && $form->isValid())
+        {
+            $hash = $encoder->encodePassword($user, $user->getPassword());
+
+            $user->setPassword($hash);
+            $user->setToken($tokenGenerator->generateToken());
+
+            $manager->persist($user);
+            $manager->flush();
+
+            $this->mailer->sendEmail($user->getEmail(), $user->getToken(), 'Confirmez votre adresse email', 'email/confirm_email.html.twig');
+
+            return $this->redirectToRoute('security_login');
+        }
+
+        return $this->render('security/registration.html.twig', [
+            'form' => $form->createView()
+        ]);
+    }
+
+    /**
+     * @Route("confirmation-compte/{token}", name="confirm_account")
+     * @param string $token
+     */
+    public function confirmAccount(string $token)
+    {
+        $user = $this->userRepository->findOneBy([
+            "token" => $token
+        ]);
+
+        if($user)
+        {
+            $user->setToken(null);
+            $user->setIsVerified(true);
+
+            $manager = $this->getDoctrine()->getManager();
+            $manager->persist($user);
+            $manager->flush();
+
+            $this->addFlash("success", "Compte actif !" );
+            return $this->redirectToRoute('trick_home');
+        }
+        else
+        {
+            $this->addFlash("error", "Ce compte n'existe pas." );
+            return $this->redirectToRoute('trick_home');
+        }
+    }
+
+}
