+++ src/Repository/UserRepository.php
@@ -64,7 +64,4 @@ class UserRepository extends ServiceEntityRepository implements PasswordUpgrader
         ;
     }
     */
-    public function findOneByEmail($email)
-    {
-    }
 }
