+++ templates/email/reset_password_email.html.twig
@@ -1,3 +1,1 @@
-salut {{ token }}
-
-<a href="{{ url('reset_password', {"token": token}) }}">Cliquez ici pour confirmer votre compte</a>
+<a href="{{ url('reset_password', {"token": token}) }}">Cliquez ici pour réinitialiser votre mot de passe</a>
