+++ templates/security/login.html.twig
@@ -10,7 +10,7 @@
    <input placeholder="Adresse email..." required name="_username" type="text" class="form-control">
    <input placeholder="Mot de passe..." required name="_password" type="password" class="form-control">
    <a href="{{ path('security_registration') }}">Pas encore inscrit ?</a>
-
+   <a href="{{ path('forgot_password') }}">Mot de passe oublié ?</a>
   </div>
 
   <div class="form-group">
