+++ templates/security/reset_password.html.twig
@@ -1,0 +1,12 @@
+{% extends 'base.html.twig' %}
+
+{% block title %}Réinitialiser mot de passe{% endblock %}
+
+{% block body %}
+    <h1>Réinitialiser mot de passe</h1>
+    {{ form_start(form) }}
+    {{ form_row(form.password, {'label': 'Mot de passe', 'attr': {'placeholder': 'Votre mot de passe...'}}) }}
+    {{ form_row(form.confirm_password, {'label': 'Confirmation mot de passe', 'attr': {'placeholder': 'Confirmez votre mot de passe...'}}) }}
+    <button type="submit" class="btn btn-success">Réinitialiser mot de passe</button>
+    {{ form_end(form) }}
+{% endblock %}
